fix: script entrypoint
