+++ src/fortress/lib/vault.py
@@ -45,13 +45,16 @@ class Vault:
         """get an entry"""
         full_path = f"{self._path}{key}"
         logging.debug(f"Vault: getting {full_path}")
-        # if key does not exist or if data is soft-deleted, it raises:
-        # hvac.exceptions.InvalidPath
-        response = self._client.secrets.kv.read_secret_version(
-            path=full_path, mount_point=self._mount_point
-        )
-        # return value of key
-        value = response["data"]["data"][key]
+        try:
+            response = self._client.secrets.kv.read_secret_version(
+                path=full_path, mount_point=self._mount_point
+            )
+            # return value of key
+            value = response["data"]["data"][key]
+        except hvac.exceptions.InvalidPath:
+            logging.debug(f"{key} either does not exist or is soft deleted.")
+            return None
+
         return value
 
     def list(self):
