--- src/load/load/context.py
@@ -1,5 +1,0 @@
-import sys
-from pathlib import Path
-
-path_root = Path(__file__).parents[2]
-sys.path.append(str(path_root))
